refactor: components 패키지 이동
