+++ app/src/androidTest/java/woowacourse/payments/CardNumberTextFieldTest.kt
@@ -14,7 +14,7 @@ import androidx.compose.ui.test.performTextInput
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-import woowacourse.payments.ui.component.CardNumberTextField
+import woowacourse.payments.ui.add.components.CardNumberTextField
 
 class CardNumberTextFieldTest {
     @get:Rule
