+++ app/src/androidTest/java/woowacourse/payments/PaymentCardTest.kt
@@ -10,7 +10,7 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import woowacourse.payments.domain.model.BankType
-import woowacourse.payments.ui.component.PaymentCard
+import woowacourse.payments.ui.cards.components.PaymentCard
 import woowacourse.payments.ui.model.PaymentCardUiModel
 import woowacourse.payments.ui.model.mapper.toUiModel
