+++ app/src/main/java/woowacourse/payments/ui/add/AddPaymentCardScreen.kt
@@ -17,13 +17,13 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import woowacourse.payments.domain.PaymentCardStore
-import woowacourse.payments.ui.component.BankSelectBottomSheet
-import woowacourse.payments.ui.component.CardNumberTextField
-import woowacourse.payments.ui.component.ExpiryTextField
-import woowacourse.payments.ui.component.NewCardTopBar
-import woowacourse.payments.ui.component.PaymentCard
-import woowacourse.payments.ui.component.PinTextField
-import woowacourse.payments.ui.component.StringTextField
+import woowacourse.payments.ui.add.components.BankSelectBottomSheet
+import woowacourse.payments.ui.add.components.CardNumberTextField
+import woowacourse.payments.ui.add.components.ExpiryTextField
+import woowacourse.payments.ui.add.components.NewCardTopBar
+import woowacourse.payments.ui.add.components.PinTextField
+import woowacourse.payments.ui.add.components.StringTextField
+import woowacourse.payments.ui.cards.components.PaymentCard
 import woowacourse.payments.ui.model.PaymentCardUiModel
 import woowacourse.payments.ui.model.mapper.toUiModelOrPlaceholder
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
