+++ app/src/main/java/woowacourse/payments/ui/add/components/BankItem.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.component
+package woowacourse.payments.ui.add.components
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
